+++ tardis/opacities/opacity_solver.py
@@ -4,7 +4,7 @@ import pandas as pd
 from tardis.opacities.opacity_state import (
     OpacityState,
 )
-from tardis.opacities.tau_sobolev import (
+from tardis.opacities.sobolev import (
     calculate_beta_sobolev,
     calculate_sobolev_line_opacity,
 )
